--- App_Code/Chat/MessageRecipient.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-[Serializable]
-public class MessageRecipient
-{
-    public MessageRecipient()
-    {
-        chatRoomIds = new List<string>();
-    }
-    public string messageRecipientId { get; set; }
-    public string messageRecipientName { get; set; }
-    public string connectionId { get; set; }
-    public List<string> chatRoomIds { get; set; }
-}
