+++ Themes/css/ChatStyle.css
@@ -7,7 +7,7 @@
 #header
 {
     padding: 20px;
-    background-color: #d7e5e4;
+    background-color: #fb8c00 !important;
     border-bottom: #5f9482 solid 2px;
     color: #1e4638;
     font-size: 24px;
@@ -15,7 +15,7 @@
 
 .submitButton
 {
-    background-color: #d7e5e4;
+    background-color: #fb8c00 !important;
     border: #5f9482 solid 1px;
     padding: 4px;
     cursor: pointer;
@@ -61,7 +61,7 @@
         font-size: 16px;
         font-weight: bold;
         padding: 8px;
-        background-color: #d7e5e4;
+        background-color: #fb8c00 !important;
         border-bottom: solid 1px #5f9482;
     }
 
@@ -102,7 +102,7 @@
                 display: block;
                 cursor: pointer;
                 padding: 4px;
-                background-color: #f9f9f9;
+                background-color: #fb8c00 !important;
                 border-bottom: solid 1px #5f9482;
             }
 
@@ -116,7 +116,7 @@
 
             .chatRoom .content .users .user:hover
             {
-                background-color: #e1e1e1;
+                background-color: #fb8c00 !important;
             }
 
     .chatRoom .messageBar
@@ -134,7 +134,7 @@
 {
     position: absolute;
     margin: 10px;
-    background-color: #ffcbcb;
+    background-color: #fb8c00 !important;
     padding: 4px;
     border: solid 1px red;
 }
@@ -150,7 +150,7 @@
     .draggable .header
     {
         cursor: move;
-        background-color: #d7e5e4;
+        background-color: #fb8c00 !important;
         border-bottom: #5f9482 solid 1px;
         color: #1e4638;
     }
